🐛 [Fix] #199 리뷰 수정사항 반영
