+++ Chalkak/Presentation/Camera/CameraViewModel.swift
@@ -98,8 +98,6 @@ class CameraViewModel: ObservableObject {
 
         loadSavedSettings()
 
-//        configure()
-
         // 뱃지 상태 초기화
         Task { @MainActor in
             updateBadgeState()
@@ -117,10 +115,6 @@ class CameraViewModel: ObservableObject {
 
     func switchCameraControls() {
         showingCameraControl.toggle()
-    }
-
-    func configure() {
-        model.requestAndCheckPermissions()
     }
 
     /// 카메라 세션 시작
